Added a Makefile target (driver-database) that downloads all the drivers from the AQUA Driver Repository

+++ src/c/drivers/irq/pit.c
@@ -17,7 +17,7 @@ void pit_handler(register_t* regs) {
 	
 }
 
-void pit_phase(int hertz) {
+void pit_phase(float hertz) {
 	int divisor = 1193182 / hertz;
 	
 	outportb(0x43, 0x36);
